feat(admin): Complete admin dashboard with full event and user management

+++ frontend/src/services/Bookings.js
@@ -3,8 +3,12 @@ import FetchErrorHandler from "./FetchErrorHandler";
 
 const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
 
-export const getAllUserBookings = FetchErrorHandler(async () => {
+export const getAllBookings = FetchErrorHandler(async () => {
     return await axios.get(`${API_BASE_URL}/bookings`, { withCredentials: true })
+});
+
+export const getAllUserBookings = FetchErrorHandler(async (userId) => {
+    return await axios.get(`${API_BASE_URL}/bookings/${userId}`, { withCredentials: true })
 });
 
 export const bookEvent = FetchErrorHandler(async (eventId) => {
@@ -12,5 +16,20 @@ export const bookEvent = FetchErrorHandler(async (eventId) => {
 });
 
 export const deleteBooking = FetchErrorHandler(async (eventId) => {
+    if (!eventId)
+        throw new Error("Event id is missing");
+
     return await axios.delete(`${API_BASE_URL}/bookings/${eventId}`,  { withCredentials: true });
+});
+
+export const updateBookingStatus = FetchErrorHandler(async (bookingId, newStatus) => {
+    if (!bookingId || !newStatus) {
+        throw new Error("Booking data is missing");
+    }
+
+    return await axios.patch(
+        `${API_BASE_URL}/bookings/${bookingId}`,
+        { newStatus },
+        { withCredentials: true }
+    );
 });
